Refactor: scraper layering
Fixed: controller오류 수정
