+++ database.js
@@ -1,6 +1,6 @@
 const { Sequelize } = require("sequelize");
 
-const sequelize = new Sequelize("im_test", "root", "1", {
+const sequelize = new Sequelize("im_test", "root", process.env.MYSQL_PASSWORD, {
   host: "localhost",
   port: 3306,
   logging: false,
