+++ routes/mygroup.js
@@ -8,7 +8,7 @@ const { isAuth } = require("../middlewares");
 const router = express.Router();
 
 router.get("/", isAuth, getPageDetails);
-router.post("/join", isAuth, joinGroup);
+router.post("/join", isAuth, joinGroup, getPageDetails);
 router.post("/update", isAuth, updateGroup, getPageDetails);
 
 module.exports = router;
