+++ scraper/allPosts/velogEvaluate.js
@@ -1,17 +1,23 @@
-module.exports = function () {
+module.exports = async () => {
   const posts = document.querySelectorAll(
     "#root > div:nth-child(2) > div:nth-child(3) > div:nth-child(4) > div:nth-child(3) > div > div"
   );
 
-  return [...posts].map((post) => {
+  let postsArray = [];
+
+  for (let post of [...posts]) {
     const link = `https://velog.io${
       post.querySelector("a")?.attributes.href.value
     }`;
     const thumbnail = post.querySelector("a > div > img")?.src;
     const title = post.querySelector("a > h2")?.innerText;
     const subtitle = post.querySelector("p")?.innerText;
-    const date = post.querySelector("div.subinfo > span")?.innerText;
 
-    return { title, subtitle, date, link, thumbnail };
-  });
+    let date = post.querySelector("div.subinfo > span")?.innerText;
+    date = await changeDateFormat(date);
+
+    postsArray = [...postsArray, { title, subtitle, date, link, thumbnail }];
+  }
+
+  return postsArray;
 };
