+++ scraper/index.js
@@ -1,0 +1,4 @@
+const getAllPosts = require("./allPosts/getAllPosts");
+const getRecentPosts = require("./recentPosts/getRecentPosts");
+
+module.exports = { getAllPosts, getRecentPosts };
