--- utils/getRandomIntInclusive.js
@@ -1,7 +1,0 @@
-module.exports = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-//곧 필요없을거같은데..?
