--- utils/shuffleArray.js
@@ -1,8 +1,0 @@
-module.exports = (array) => {
-  for (let i = array.length - 1; i > 0; i--) {
-    let j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-};
-
-//곧 필요없을거같은데..?
